:construction: More draw_windows() progress
Continued work on drawing multiple windows to the same screen. Multiple windows can be drawn but a similar issue to single window drawing arises in that although all rows of the terminal screen are drawn, the bottom several are not rendered until the terminal is updated via a scroll for example.